# small perf improvements to TriangulateSpade